chore(api): implementation of API testing (#104)
