+++ functions/src/index.ts
@@ -1,2 +1,6 @@
-export * from './google/places';
-export * from './sendgrid/mail';
+/* Google API */
+export * from './google/places-find';
+export * from './google/places-search';
+
+/* SendGrid API */
+export * from './sendgrid/contact-email';
