+++ hosting/src/app/events/create/event-create.component.ts
@@ -82,7 +82,7 @@ export class EventCreateComponent implements OnInit {
             .searchPlace(searchValue, 'establishment', this.googleSession)
             .pipe(
                 map((result) => {
-                    this.googleSession = result.uid;
+                    this.googleSession = result.sessionId;
                     return result.places.map((place) => ({ id: place.id, text: place.description }));
                 }),
             );
